requirements_server.txt
absl-py==2.4.0
action-msgs==1.2.2
action-tutorials-interfaces==0.20.6
action-tutorials-py==0.20.6
actionlib-msgs==4.9.0
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
anyio==4.12.1
builtin-interfaces==1.2.2
certifi==2026.1.4
click==8.3.1
composition-interfaces==1.2.2
contourpy==1.3.3
cv-bridge==3.2.1
cycler==0.12.1
demo-nodes-py==0.20.6
diagnostic-msgs==4.9.0
domain-coordinator==0.10.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.4
examples-rclpy-minimal-action-client==0.15.4
examples-rclpy-minimal-action-server==0.15.4
examples-rclpy-minimal-client==0.15.4
examples-rclpy-minimal-publisher==0.15.4
examples-rclpy-minimal-service==0.15.4
examples-rclpy-minimal-subscriber==0.15.4
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
geometry-msgs==4.9.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
idna==3.11
image-geometry==3.2.1
interactive-markers==2.3.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
laser-geometry==2.4.1
launch==1.0.11
launch-ros==0.19.12
launch-testing==1.0.11
launch-testing-ros==0.19.12
launch-xml==1.0.11
launch-yaml==1.0.11
lifecycle-msgs==1.2.2
logging-demo==0.20.6
map-msgs==2.1.0
Markdown==3.10.1
MarkupSafe==2.1.5
matplotlib==3.10.8
message-filters==4.3.11
mpmath==1.3.0
nav-msgs==4.9.0
networkx==3.6.1
numpy==2.4.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.13.0.90
osrf-pycommon==2.1.6
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pandas==3.0.0
pcl-msgs==1.0.0
pendulum-msgs==0.20.6
pillow==12.1.0
protobuf==6.33.4
psutil==7.2.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-qt-binding==1.1.3
PyYAML==6.0.3
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.6
rcl-interfaces==1.2.2
rclpy==3.3.17
rcutils==5.1.7
resource-retriever==3.1.3
rmw-dds-common==1.6.0
ros2action==0.18.14
ros2bag==0.15.15
ros2cli==0.18.14
ros2component==0.18.14
ros2doctor==0.18.14
ros2interface==0.18.14
ros2launch==0.19.12
ros2lifecycle==0.18.14
ros2multicast==0.18.14
ros2node==0.18.14
ros2param==0.18.14
ros2pkg==0.18.14
ros2run==0.18.14
ros2service==0.18.14
ros2topic==0.18.14
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.7
rosidl-cli==3.1.7
rosidl-cmake==3.1.7
rosidl-generator-c==3.1.7
rosidl-generator-cpp==3.1.7
rosidl-generator-py==0.14.6
rosidl-parser==3.1.7
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.3
rosidl-typesupport-fastrtps-cpp==2.2.3
rosidl-typesupport-introspection-c==3.1.7
rosidl-typesupport-introspection-cpp==3.1.7
rpyutils==0.2.2
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.8
rqt-gui-py==1.1.8
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.8
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
shape-msgs==4.9.0
shellingham==1.5.4
six==1.17.0
sros2==0.10.7
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
sympy==1.13.1
teleop-twist-keyboard==2.4.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tf2-geometry-msgs==0.25.16
tf2-kdl==0.25.16
tf2-msgs==0.25.16
tf2-py==0.25.16
tf2-ros-py==0.25.16
tf2-sensor-msgs==0.25.16
tf2-tools==0.25.16
threadpoolctl==3.6.0
timm==1.0.24
topic-monitor==0.20.6
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
trajectory-msgs==4.9.0
triton==3.2.0
turtlesim==1.4.3
typer-slim==0.21.1
typing_extensions==4.15.0
unique-identifier-msgs==2.2.1
visualization-msgs==4.9.0
Werkzeug==3.1.5
